+++ .eslintrc.js
@@ -6,4 +6,10 @@ module.exports = require('@hugsmidjan/hxmstyle')({
     node: true,
     es2020: true,
   },
+  rules: {
+    // Require file extensions for all local imports.
+    // Otherwise tsc emits ESM modules that don't work with Next.js
+    // (and possibly other bundlers)
+    'import/extensions': ['error', 'ignorePackages'],
+  },
 });